requirements.txt
awsebcli==3.4.7
backports.ssl-match-hostname==3.4.0.2
cement==2.4.0
Django==1.8.3
django-tenant-schemas==1.5.5
djangorestframework==3.2.0
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
jmespath==0.7.1
pathspec==0.3.3
psycopg2==2.6.1
python-dateutil==2.4.2
PyYAML==3.11
requests==2.6.2
six==1.9.0
texttable==0.8.3
websocket-client==0.32.0
wheel==0.24.0
